requirements.txt
#Django>=3.0.6,<3.1.0
Django>=2.2.0,<2.3.0
django-filter>=2.2.0,<2.3.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.7.5,<2.8.0
flake8>=3.8.2,<3.9.0
# required by social_django
social-auth-app-django>=3.1.0,<3.2.0
# required by extra app xadmin
django-crispy-forms>=1.9.0,<1.10.0
django-formtools==2.2
django-import-export>=2.0.2,<2.1.0
future>=0.18.2,<0.19.0
httplib2>=0.17.3,<0.18.0
six>=1.14.0,<1.15.0
# required by utils
requests>=2.23.0,<2.24.0
